Blocks streaming API (#185)
* Blocks streaming API.

* Deprecated methods eliminated.

* Endpoint descr updated.

* Bump version.

+++ modules/explorer-core/src/main/scala/org/ergoplatform/explorer/db/queries/BlockInfoQuerySet.scala
@@ -71,7 +71,7 @@ object BlockInfoQuerySet extends QuerySet {
          |""".stripMargin.query[BlockStats]
 
   def getManyExtendedMain(
-    offset: Int,
+    offset: Long,
     limit: Int,
     ordering: OrderingString,
     orderBy: String
